Some minor fixes
These updates fixed a couple of minor problems with the game. The game as submitted before is playable but could get stuck if you clicked on a card that was not the right suit. Also, if you started the bidding, when the other three players had submitted their bid, it would tell you that it is your turn to bid instead of your turn to play. This did not affect the game just gave the wrong message.